supplier view component

+++ src/app/components/dashboard/sidebar/sidebar.component.ts
@@ -5,6 +5,9 @@ import { map, shareReplay } from 'rxjs/operators';
 import { AuthService } from 'src/app/shared/services/auth.service';
 import { Router } from '@angular/router';
 import { UsersService } from 'src/app/shared/services/users.service';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { AlertComponent } from '../alert/alert.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-sidebar',
@@ -22,10 +25,13 @@ export class SidebarComponent {
 
   constructor(
     private authService: AuthService,
+    public dialog: MatDialog,
+    private _snackBar: MatSnackBar,
     public usersService: UsersService,
     private router: Router
   ) { }
   user$ = this.usersService.currentUserProfile$;
+
   logOut() {
     // this.authService.logout().subscribe(async user => {
     //   if (user) {
@@ -34,12 +40,41 @@ export class SidebarComponent {
     //     location.reload();
     //   }
     // })
-    this.authService.logout().subscribe(() => {
-      localStorage.clear();
-      this.router.navigate(['/loginuser']);
-      location.reload();
-    });
+
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = false;
+    dialogConfig.autoFocus = true;
+    dialogConfig.data = {
+      tittle: 'Alert!',
+      action: 'LogOut'
+    }
+    const dialogRef = this.dialog.open(AlertComponent, dialogConfig);
+
+    dialogRef.afterClosed().subscribe(data => {
+      if (data) {
+        // this.supplierService.deleteSupplier(row);
+        // console.log(row);
+        this.authService.logout().subscribe(() => {
+          localStorage.clear();
+          this.router.navigate(['/loginuser']);
+          location.reload();
+        });
+
+        this.openSnackBar("Supplier deleted Successfully!", "OK")
+      }
+    })
 
 
+
+
+
+
+
+
+
+  }
+
+  openSnackBar(message: string, action: string) {
+    this._snackBar.open(message, action);
   }
 }
